Fix Makefile INCLUDEs and JavaFX maven dependency
The Makefile was not passing the INCLUDE
to the recipe for compiling the native
agent, and the compilation was therefore
failing to find the requisite spdlog
headers.

Furthermore, JavaFX is required to build
the Java portion of JCoz. When the tool
was first built, there was no maven plugin
for JavaFX, and it was a nightmare to get
maven to point to the right jar file for
use when building the JCoz jar output.
This change updates pom to use the
now-available JavaFX plugin.