Remove vernetperronllc from JCoz
Fixes #38

Once upon a time, it was under consideration
to try and monetize JCoz. This is no longer
the case, and JCoz is free and open source
under the permissions afforded by the GPL.
This change therefore removes references to
vernetperronllc, and restructures the layout
of the Java portion of JCoz accordingly.

Testing Done: Verified that java and tests
make targets can still be built, and that
docker image can still be built.