Remove JavaFX / UI client from JCoz
Fixes #39

Up until this change, JCoz supported
both UI and CLI clients for profiling
a given JVM process. The UI client
could render the results of a profile
in real time, whereas the CLI client
simply supported configuring and
connecting to a profiling session.

In addition to JavaFX not really being
used by most Java authors for rendering
GUI, it is one of the more painful parts
of building and porting JCoz to other
platforms, and adds a lot of complexity
to the codebase.

This change therefore removes JavaFX, and
all associated logic, from JCoz.

Testing Done: Successfully built project
using both Dockerfile, and Dockerfile.fedora