Remove unused ParseOptions logic from native agent
In JCoz' early days, there was no Java
component, and profiling a JVM process
simply involved invoking java on the
process with the native agent, and a
set of passed parameters. Now that
profiling parameters such as progress
point class/line number are passed down
to the native agent through JNI calls
from the JCoz Java service, the user
need no longer configure where the
progress point is set via passing a
long continuous argument string as
part of the java nativeagent parameter.

As a first step in the direction towards
documenting and refactoring the native
Profiler class (#56), this commit
removes the logic from the nativeagent
to parse and configure nativeagent
parameters from the user.