Refactor code related with several scopes support in entry.cc
Methods that check prefix were renamed to be clearer in what they are doing.

`sizeof(char)` was replaced with just `1` for the same reason. Explanation why this addition is needed at all added as a comment.

Also, predicate as parameter of `contains_class_fqn_prefix` is obsolete; now it takes `char*` instead.