Update License to Apache V2
Over email, we've reached out ot all of the existing contributors to JCoz to
get their permission to change the license to Apache v2, and all except one has
agreed (with one not responding). All commits by the one author from whom we
didn't receive a response have been reverted.

With that, this change updates the license for JCoz from GPLv3, to Apache 2.

Signed-off-by: David Vernet <void@manifault.com>